pyproject.toml
[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "airtrackrelay"
version = "1.0.4"
description = "UDP GPS Tracker Relay"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
keywords = ["GPS", "UDP", "tracking"]
authors = [
    {email = "ndf-zz@6-v.org", name = "Nathan Fraser"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Other/Nonlisted Topic",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "metarace",
    "pycryptodomex",
]

[project.urls]
homepage = "https://github.com/ndf-zz/airtrackrelay"

[project.scripts]
airtrackrelay = "airtrackrelay:main"
